# docs/dist/openapi.yaml
swagger: "2.0"
info:
    version: 1.0.0
    title: GetYTStatsAPI
    description: '
        <p>Проект представляет собой API для получения статистики видео, подсчёта общего количества просмотров, и экспорта в CSV файл.</p>

        <p>По вопросам работы с API обращайтесь <a href="https://t.me/ra1ph_d">сюда</a>.</p>
        <p>Если вы обнаружили ошибку или у вас есть предложения по улучшению, создавайте issue на <a href="https://github.com/ra1phdd/GetYTStatsAPI">GitHub</a>.</p>

        <h2>Как работать с документацией?</h2>
        <p>Для отправки запросов к API нажмите на кнопку Try it out и заполните необходимые поля для составления нужного фильтра.
        После заполнения полей нажмите на кнопку Execute и получите ответ от API и пример запроса.</p>'
host: 91.215.21.236:8089
basePath: /
paths:
    /get_command:
        get:
            summary: Получение статистики видео по идентификатору канала, ключевому слову для рекламы, дате начала и максимальному количеству результатов.
            description: Этот эндпоинт возвращает команду для вставки в Excel (Google Sheets)
            operationId: getCommandHandler
            tags:
                - stats
            parameters:
                - name: channel_id
                  in: query
                  required: true
                  type: string
                  description: ID канала на YouTube (пример, UCNPUUqi4kqjeaScWtsvfyvw)
                - name: ad_word
                  in: query
                  required: true
                  type: string
                  description: Ключевое слово для рекламы из описания (пример, LOLZTEAM)
                - name: start_date
                  in: query
                  required: true
                  type: string
                  format: string
                  description: Дата старта рекламы в формате YYYY-MM-DD (пример, 2024-07-14)
                - name: max_results
                  in: query
                  required: false
                  type: integer
                  description: Максимальное количество видео, по которым сервис будет осуществлять подсчёт статистики
            responses:
                '200':
                    description: Команда для вставки в Excel (Google Sheets)
                    schema:
                        type: string
                '500':
                    description: Внутренняя ошибка сервера
                    schema:
                        type: string